Task 4.21 done, added token check for blog removal

+++ controllers/blogs.js
@@ -32,6 +32,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blogToBeRemoved = await Blog.findById(request.params.id)
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
+  console.log(`User Id from request: ${decodedToken.id}`)
+  console.log(`User Id from blog: ${blogToBeRemoved.user.toString()}`)
+  
+  if (blogToBeRemoved.user.toString() !== decodedToken.id ) {
+    return response.status(401).json({ error: 'not authorized: only the creator can delete a blog' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
